[Refactor] Refactor CCE get job
