+++ openstack/cce/v3/nodes/urls.go
@@ -5,7 +5,3 @@ import golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 func rootURL(c *golangsdk.ServiceClient, clusterid string) string {
 	return c.ServiceURL("clusters", clusterid, "nodes")
 }
-
-func getJobURL(c *golangsdk.ServiceClient, jobid string) string {
-	return c.ServiceURL("jobs", jobid)
-}
